✅ Membatasi Hak Akses User Multi Role

+++ routes/web.php
@@ -20,7 +20,7 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
     return view('home');
-});
+})->middleware('auth');
 
 Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
 Route::post('/login', [AuthController::class, 'authenticating'])->middleware('guest');
@@ -30,15 +30,15 @@ Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
 
 
 Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
-Route::get('/student-add', [StudentController::class, 'create'])->middleware('auth');
-Route::post('/student-store', [StudentController::class, 'store'])->middleware('auth');
-Route::get('/student-edit/{id}', [StudentController::class, 'edit'])->middleware('auth');
-Route::put('/student-update/{id}', [StudentController::class, 'update'])->middleware('auth');
-Route::get('/student-detail/{id}', [StudentController::class, 'show'])->middleware('auth');
-Route::get('/student-delete/{id}', [StudentController::class, 'delete'])->middleware('auth');
-Route::delete('/student-destroy/{id}', [StudentController::class, 'destroy'])->middleware('auth');
-Route::get('/students-showdeleted', [StudentController::class, 'showdeleted'])->middleware('auth');
-Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware('auth');
+Route::get('/student-add', [StudentController::class, 'create'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::post('/student-store', [StudentController::class, 'store'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-edit/{id}', [StudentController::class, 'edit'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::put('/student-update/{id}', [StudentController::class, 'update'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-detail/{id}', [StudentController::class, 'show'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-delete/{id}', [StudentController::class, 'delete'])->middleware(['auth', 'must-admin']);
+Route::delete('/student-destroy/{id}', [StudentController::class, 'destroy'])->middleware(['auth', 'must-admin']);
+Route::get('/students-showdeleted', [StudentController::class, 'showdeleted'])->middleware(['auth', 'must-admin']);
+Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware(['auth', 'must-admin']);
 
 Route::get('/class', [ClassController::class, 'index'])->middleware('auth');
 Route::get('/class-add', [ClassController::class, 'create'])->middleware('auth');
